Improvements and Error Handling
