+++ app/controllers/errors_controller.rb
@@ -1,0 +1,9 @@
+class ErrorsController < ApplicationController
+  def not_found
+    render_error("Not Found", :not_found)
+  end
+
+  def internal_server_error
+    render_error("Something went wrong", :internal_server_error)
+  end
+end
