+++ db/seeds.rb
@@ -5,3 +5,5 @@
 #
 #   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
 #   Mayor.create(name: 'Emanuel', city: cities.first)
+
+DataEncryptingKey.generate!(primary: true) if DataEncryptingKey.primary.nil?
